currency conversion alpha v1
