+++ ff_data_server/src/request_handlers.rs
@@ -17,7 +17,7 @@ use tokio::sync::mpsc::{Receiver, Sender};
 use tokio::time::timeout;
 use tokio_rustls::server::TlsStream;
 use crate::server_features::database::DATA_STORAGE;
-use crate::server_side_brokerage::{account_info_response, accounts_response, commission_info_response, intraday_margin_required_response, overnight_margin_required_response, live_market_order, symbol_info_response, symbol_names_response, live_enter_long, live_exit_long, live_exit_short, live_enter_short, other_orders, cancel_order, flatten_all_for, update_order, cancel_orders_on_account};
+use crate::server_side_brokerage::{account_info_response, accounts_response, commission_info_response, intraday_margin_required_response, overnight_margin_required_response, live_market_order, symbol_info_response, symbol_names_response, live_enter_long, live_exit_long, live_exit_short, live_enter_short, other_orders, cancel_order, flatten_all_for, update_order, cancel_orders_on_account, exchange_rate_response};
 use crate::server_side_datavendor::{base_data_types_response, decimal_accuracy_response, markets_response, resolutions_response, symbols_response, tick_size_response};
 use ff_standard_lib::standardized_types::enums::StrategyMode;
 use ff_standard_lib::standardized_types::orders::{Order, OrderRequest, OrderType, OrderUpdateEvent};
@@ -92,7 +92,7 @@ pub async fn manage_async_requests(
                     continue;
                 }
             };
-            println!("{:?}", request);
+            //println!("{:?}", request);
 
             let stream_name = stream_name.clone();
             let mode = strategy_mode.clone();
@@ -102,6 +102,18 @@ pub async fn manage_async_requests(
                 // Handle the request and generate a response
                 match request {
                     DataServerRequest::Register(_) => {},
+                    DataServerRequest::ExchangeRate {
+                        callback_id,
+                        from_currency,
+                        to_currency,
+                        date_time_string,
+                        data_vendor
+                    } => {
+                        handle_callback(
+                            ||  exchange_rate_response(mode, from_currency, to_currency, date_time_string, data_vendor, callback_id),
+                            sender.clone()
+                        ).await
+                    }
                     DataServerRequest::DecimalAccuracy {
                         data_vendor,
                         callback_id,
@@ -287,7 +299,6 @@ pub async fn manage_async_requests(
                     DataServerRequest::RegisterStreamer { .. } => {
                         //no need to handle here
                     }
-                    DataServerRequest::Rates { .. } => {}
                 }
             });
         }
