+++ ff_standard_lib/src/standardized_types/datavendor_enum.rs
@@ -5,7 +5,7 @@ use serde_derive::{Deserialize, Serialize};
 use strum_macros::EnumIter;
 use crate::messages::data_server_messaging::FundForgeError;
 
-#[derive(Serialize, Deserialize, Clone, Eq, Serialize_rkyv, Deserialize_rkyv, Archive, PartialEq, Debug, Hash, PartialOrd, Ord, EnumIter)]
+#[derive(Serialize, Deserialize, Clone, Eq, Serialize_rkyv, Deserialize_rkyv, Archive, PartialEq, Debug, Hash, PartialOrd, Ord, EnumIter, Copy)]
 #[archive(compare(PartialEq), check_bytes)]
 #[archive_attr(derive(Debug))]
 /// A `DataVendor` enum is a company that provides the data that is used to feed the algorithm.
