+++ ff_standard_lib/src/standardized_types/mod.rs
@@ -13,5 +13,3 @@ pub mod time_slices;
 pub mod position;
 pub mod books;
 pub mod accounts;
-pub mod futures_products;
-pub mod market_maps;
