+++ ff_standard_lib/src/strategies/client_features/mod.rs
@@ -9,3 +9,4 @@ pub(crate) mod live_subscriptions;
 pub(crate) mod request_handler;
 mod response_handler;
 mod live_data_receiver;
+pub mod other_requests;
