+++ ff_standard_lib/src/strategies/consolidators/consolidator_enum.rs
@@ -6,10 +6,10 @@ use crate::standardized_types::enums::{MarketType, StrategyMode};
 use crate::standardized_types::rolling_window::RollingWindow;
 use crate::standardized_types::subscriptions::{filter_resolutions, CandleType, DataSubscription};
 use chrono::{DateTime, Datelike, Duration, Utc, Weekday};
+use crate::product_maps::rithmic::maps::extract_symbol_from_contract;
 use crate::standardized_types::base_data::base_data_type::BaseDataType;
 use crate::standardized_types::base_data::history::get_historical_data;
 use crate::standardized_types::resolution::Resolution;
-use crate::standardized_types::futures_products::extract_symbol_from_contract;
 
 pub enum ConsolidatorEnum {
     Count(CountConsolidator),
